+++ TPie/Helpers/WotsitHelper.cs
@@ -58,7 +58,10 @@ namespace TPie.Helpers
         public void Update()
         {
             _map.Clear();
-            UnregisterAll();
+            if (!UnregisterAll())
+            {
+                return; 
+            }
 
             List<Ring> rings = Plugin.Settings.Rings;
 
@@ -85,9 +88,17 @@ namespace TPie.Helpers
             }
         }
 
-        public void UnregisterAll()
+        public bool UnregisterAll()
         {
-            _unregisterAll.InvokeFunc(Plugin.PluginInterface.InternalName);
+            try
+            {
+                _unregisterAll.InvokeFunc(Plugin.PluginInterface.InternalName);
+                return true;
+            } 
+            catch
+            {
+                return false;
+            }
         }
     }
 }