+++ TPie/Plugin.cs
@@ -87,7 +87,7 @@ namespace TPie
                 AssemblyLocation = Assembly.GetExecutingAssembly().Location;
             }
 
-            Version = Assembly.GetExecutingAssembly().GetName().Version?.ToString() ?? "1.7.1.0";
+            Version = Assembly.GetExecutingAssembly().GetName().Version?.ToString() ?? "1.7.1.1";
 
             Framework.Update += Update;
             UiBuilder.Draw += Draw;